requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
filelock==3.6.0
huggingface-hub==0.5.1
idna==3.3
joblib==1.1.0
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pymongo==4.1.1
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
sacremoses==0.0.53
six==1.16.0
tokenizers==0.12.1
torch==1.11.0
tqdm==4.64.0
transformers==4.18.0
tweepy==4.8.0
typing_extensions==4.2.0
urllib3==1.26.9
